feat: add remaining duration info
Cool new thing. However, the UI will flicker every second as it has to
be completely redrawn. Annoying, but haven't found a solution for it
yet.

Also, m4a support is gone. It was never actually there, I just mistook
that miniaudio had it by default. It is implementable, but I'm not gonna
bother.